+++ tests/commands/test_regions.py
@@ -22,6 +22,7 @@ from unittest import mock
 import pytest
 
 from croud.api import Client, RequestMethod
+from croud.printer import RawFormatPrinter
 from tests.util import assert_rest, call_command
 
 
@@ -94,11 +95,50 @@ def test_regions_create_mandatory_params(mock_request):
     )
 
 
+@mock.patch.object(RawFormatPrinter, "print_rows")
+@mock.patch.object(
+    Client,
+    "request",
+    side_effect=[({"name": "region-name"}, None), ({"token": "my-token"}, None)],
+)
+def test_regions_create_install_command(mock_request, mock_raw_printer):
+    call_command(
+        "croud",
+        "regions",
+        "create",
+        "--description",
+        "region-description",
+        "--provider",
+        "EDGE",
+        "--yes",
+    )
+
+    mock_request.call_args_list[0].assert_called_with(
+        RequestMethod.POST,
+        "/api/v2/regions/",
+        body={"description": "region-description", "provider": "EDGE"},
+        params=None,
+    )
+    mock_request.call_args_list[1].assert_called_with(
+        RequestMethod.GET, "/api/v2/regions/region-name/install-token/", params=None
+    )
+    assert any(
+        "$ bash <( wget -qO- https://console.cratedb-dev.cloud/edge/region-name/cratedb-cloud-edge.sh) my-token"  # noqa
+        in item
+        for item in mock_raw_printer.call_args[0][0]
+    )
+
+
 @mock.patch.object(Client, "request", return_value=({}, None))
 def test_regions_create_missing_description(mock_request, capsys):
     with pytest.raises(SystemExit):
         call_command(
-            "croud", "regions", "create", "--provider", "EDGE", "--yes",
+            "croud",
+            "regions",
+            "create",
+            "--provider",
+            "EDGE",
+            "--yes",
         )
     mock_request.assert_not_called()
     _, err_output = capsys.readouterr()