+++ CHANGES.rst
@@ -5,6 +5,8 @@ Changes for croud
 Unreleased
 ==========
 
+- Added a new type of printer ``raw`` that simply outputs the rows as given.
+
 - Added new subcommand ``regions generate-deployment-manifest`` to fetch a deployment
   manifest for an edge region.